Move parser into namespace to prevent collisions
