+++ lib/simple_xml/parser.rb
