+++ lib/simplexml_parser.rb
@@ -12,5 +12,5 @@ require_relative 'model/data_criteria'
 require_relative 'model/population_criteria'
 require_relative 'model/precondition'
 require_relative 'model/sub_tree'
-require_relative 'parser'
+require_relative 'simple_xml/parser'
